Drop use of org.kohsuke github-api, in favour of home-grown
Took a while to decide to make this change, but finally got here!

The old github-api library has a few issues for my use, which I'm aiming
to address:

* Blocking calls - Scala prefers Futures
* Limited analytics for quota usage: see `RateLimit` case class
* No case classes, because it's not Scala
* The value classes have icky state in them (a ref back to the connection)
* Understandably, sometimes the particular call I want isn't there

There was also a confusing & serious issue around http caching, which I
probably could have solved while staying with the org.kohsuke github-api,
but I blatently just wanted an excuse to do this.