Tolerate intermittent List wrappers used in JSON responses from the GitHub API
Surprisingly the GitHub API sometimes, for a short period of time, will return
a response with a surplus List wrapper (`[]`) around responses that are always
supposed to return a single entity - I've observed this particularly for the
'Get a reference' endpoint:

https://docs.github.com/en/rest/git/refs?apiVersion=2022-11-28#get-a-reference

```
GET /repos/{owner}/{repo}/git/ref/{ref}
```

So for a short time the response returned will start like this:

```
[{"ref":"refs/heads/feature-1", ... }]
```

...and then will settle on the expected response, which is identical except for
the removed list wrapper:

```
{"ref":"refs/heads/feature-1", ... }
```

I've observed this happening when the entity has been recently created - if you
repeatedly make the request, after a few seconds the response settles to the
normal response format, without the list-wrapper.

One way to handle this is simply performing retries, but it is also possible to
tweak the parser to discard the outer wrapper if initial parsing fails - and then
just parse the single element contained in the list. That's what this change does.